requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow





- Создать записи в БД

- Создать вьюсет для продуктов и
сделать следующие разрешения:
создавать, редактировать и удалять
 продукт может только админ

Залить проект в Github